refactor(tools): split runtime config narrowing seams
