+++ crates/app/src/tools/runtime_config.rs
@@ -14,15 +14,17 @@ use crate::conversation::{
     ConstrainedSubagentContractView, ConstrainedSubagentControlScope, ConstrainedSubagentMode,
     ConstrainedSubagentRole, ConstrainedSubagentRuntimeBinding,
 };
-#[cfg(test)]
-use crate::conversation::{
-    ConstrainedSubagentExecution, ConstrainedSubagentIdentity, ConstrainedSubagentProfile,
-};
 #[cfg(feature = "feishu-integration")]
 use crate::secrets::has_configured_secret_ref;
 
+#[path = "runtime_config_narrowing.rs"]
+mod runtime_narrowing;
 #[path = "runtime_config_web_search.rs"]
 mod web_search_runtime;
+pub(crate) use runtime_narrowing::merge_runtime_narrowing_sources;
+pub use runtime_narrowing::{
+    BrowserRuntimeNarrowing, ToolRuntimeNarrowing, WebFetchRuntimeNarrowing,
+};
 pub use web_search_runtime::WebSearchRuntimePolicy;
 
 pub const SKILLS_ENABLED_ENV: &str = "LOONG_SKILLS_ENABLED";
@@ -31,191 +33,6 @@ pub const SKILLS_ALLOWED_DOMAINS_ENV: &str = "LOONG_SKILLS_ALLOWED_DOMAINS";
 pub const SKILLS_BLOCKED_DOMAINS_ENV: &str = "LOONG_SKILLS_BLOCKED_DOMAINS";
 pub const SKILLS_INSTALL_ROOT_ENV: &str = "LOONG_SKILLS_INSTALL_ROOT";
 pub const SKILLS_AUTO_EXPOSE_INSTALLED_ENV: &str = "LOONG_SKILLS_AUTO_EXPOSE_INSTALLED";
-fn bool_is_false(value: &bool) -> bool {
-    !*value
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct BrowserRuntimeNarrowing {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_sessions: Option<usize>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_links: Option<usize>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_text_chars: Option<usize>,
-}
-
-impl BrowserRuntimeNarrowing {
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.max_sessions.is_none() && self.max_links.is_none() && self.max_text_chars.is_none()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct WebFetchRuntimeNarrowing {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub allow_private_hosts: Option<bool>,
-    #[serde(default, skip_serializing_if = "bool_is_false")]
-    pub enforce_allowed_domains: bool,
-    #[serde(default, skip_serializing_if = "BTreeSet::is_empty")]
-    pub allowed_domains: BTreeSet<String>,
-    #[serde(default, skip_serializing_if = "BTreeSet::is_empty")]
-    pub blocked_domains: BTreeSet<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub timeout_seconds: Option<u64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_bytes: Option<usize>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_redirects: Option<usize>,
-}
-
-impl WebFetchRuntimeNarrowing {
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.allow_private_hosts.is_none()
-            && !self.enforces_allowed_domains()
-            && self.allowed_domains.is_empty()
-            && self.blocked_domains.is_empty()
-            && self.timeout_seconds.is_none()
-            && self.max_bytes.is_none()
-            && self.max_redirects.is_none()
-    }
-
-    #[must_use]
-    pub fn enforces_allowed_domains(&self) -> bool {
-        self.enforce_allowed_domains || !self.allowed_domains.is_empty()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct ToolRuntimeNarrowing {
-    #[serde(default)]
-    pub browser: BrowserRuntimeNarrowing,
-    #[serde(default)]
-    pub web_fetch: WebFetchRuntimeNarrowing,
-}
-
-impl ToolRuntimeNarrowing {
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.browser.is_empty() && self.web_fetch.is_empty()
-    }
-
-    #[must_use]
-    pub fn intersect(&self, other: &Self) -> Self {
-        if self.is_empty() {
-            return other.clone();
-        }
-        if other.is_empty() {
-            return self.clone();
-        }
-
-        let browser = BrowserRuntimeNarrowing {
-            max_sessions: min_optional_limit(self.browser.max_sessions, other.browser.max_sessions),
-            max_links: min_optional_limit(self.browser.max_links, other.browser.max_links),
-            max_text_chars: min_optional_limit(
-                self.browser.max_text_chars,
-                other.browser.max_text_chars,
-            ),
-        };
-
-        let left_enforces_allowed_domains = self.web_fetch.enforces_allowed_domains();
-        let right_enforces_allowed_domains = other.web_fetch.enforces_allowed_domains();
-        let mut allowed_domains = BTreeSet::new();
-        let mut enforce_allowed_domains = false;
-
-        if left_enforces_allowed_domains && right_enforces_allowed_domains {
-            enforce_allowed_domains = true;
-            let left_is_deny_all = self.web_fetch.allowed_domains.is_empty();
-            let right_is_deny_all = other.web_fetch.allowed_domains.is_empty();
-            if !left_is_deny_all && !right_is_deny_all {
-                allowed_domains = self
-                    .web_fetch
-                    .allowed_domains
-                    .intersection(&other.web_fetch.allowed_domains)
-                    .cloned()
-                    .collect();
-            }
-        } else if left_enforces_allowed_domains {
-            enforce_allowed_domains = true;
-            allowed_domains = self.web_fetch.allowed_domains.clone();
-        } else if right_enforces_allowed_domains {
-            enforce_allowed_domains = true;
-            allowed_domains = other.web_fetch.allowed_domains.clone();
-        }
-
-        let allow_private_hosts = intersect_private_host_setting(
-            self.web_fetch.allow_private_hosts,
-            other.web_fetch.allow_private_hosts,
-        );
-
-        let blocked_domains = self
-            .web_fetch
-            .blocked_domains
-            .union(&other.web_fetch.blocked_domains)
-            .cloned()
-            .collect();
-
-        let web_fetch = WebFetchRuntimeNarrowing {
-            allow_private_hosts,
-            enforce_allowed_domains,
-            allowed_domains,
-            blocked_domains,
-            timeout_seconds: min_optional_limit(
-                self.web_fetch.timeout_seconds,
-                other.web_fetch.timeout_seconds,
-            ),
-            max_bytes: min_optional_limit(self.web_fetch.max_bytes, other.web_fetch.max_bytes),
-            max_redirects: min_optional_limit(
-                self.web_fetch.max_redirects,
-                other.web_fetch.max_redirects,
-            ),
-        };
-
-        Self { browser, web_fetch }
-    }
-}
-
-fn min_optional_limit<T>(left: Option<T>, right: Option<T>) -> Option<T>
-where
-    T: Ord + Copy,
-{
-    match (left, right) {
-        (Some(left), Some(right)) => Some(left.min(right)),
-        (Some(left), None) => Some(left),
-        (None, Some(right)) => Some(right),
-        (None, None) => None,
-    }
-}
-
-fn intersect_private_host_setting(left: Option<bool>, right: Option<bool>) -> Option<bool> {
-    if left == Some(false) || right == Some(false) {
-        return Some(false);
-    }
-    None
-}
-
-pub(crate) fn merge_runtime_narrowing_sources(
-    primary_runtime_narrowing: Option<ToolRuntimeNarrowing>,
-    secondary_runtime_narrowing: Option<ToolRuntimeNarrowing>,
-) -> Option<ToolRuntimeNarrowing> {
-    let primary_runtime_narrowing =
-        primary_runtime_narrowing.filter(|runtime_narrowing| !runtime_narrowing.is_empty());
-    let secondary_runtime_narrowing =
-        secondary_runtime_narrowing.filter(|runtime_narrowing| !runtime_narrowing.is_empty());
-
-    match (primary_runtime_narrowing, secondary_runtime_narrowing) {
-        (Some(primary_runtime_narrowing), Some(secondary_runtime_narrowing)) => {
-            let merged_runtime_narrowing =
-                primary_runtime_narrowing.intersect(&secondary_runtime_narrowing);
-            Some(merged_runtime_narrowing)
-        }
-        (Some(primary_runtime_narrowing), None) => Some(primary_runtime_narrowing),
-        (None, Some(secondary_runtime_narrowing)) => Some(secondary_runtime_narrowing),
-        (None, None) => None,
-    }
-}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct SkillsRuntimePolicy {
@@ -2812,252 +2629,6 @@ mod tests {
         );
     }
 
-    #[test]
-    fn delegate_child_prompt_summary_returns_none_when_narrowing_is_empty() {
-        assert_eq!(
-            ToolRuntimeConfig::default().delegate_child_prompt_summary(None),
-            None
-        );
-    }
-
-    #[test]
-    fn delegate_child_prompt_summary_is_effective_stable_and_sparse() {
-        let base = ToolRuntimeConfig {
-            browser: BrowserRuntimePolicy {
-                enabled: true,
-                max_sessions: 1,
-                max_links: 4,
-                max_text_chars: 1_024,
-            },
-            web_fetch: WebFetchRuntimePolicy {
-                enabled: true,
-                allow_private_hosts: false,
-                enforce_allowed_domains: true,
-                allowed_domains: BTreeSet::from(["api.example.com".to_owned()]),
-                blocked_domains: BTreeSet::from(["base-block.example.com".to_owned()]),
-                timeout_seconds: 3,
-                max_bytes: 2_048,
-                max_redirects: 1,
-            },
-            ..ToolRuntimeConfig::default()
-        };
-        let narrowing = ToolRuntimeNarrowing {
-            browser: BrowserRuntimeNarrowing {
-                max_sessions: Some(8),
-                max_links: Some(8),
-                max_text_chars: Some(512),
-            },
-            web_fetch: WebFetchRuntimeNarrowing {
-                allow_private_hosts: Some(true),
-                enforce_allowed_domains: false,
-                allowed_domains: BTreeSet::from(["docs.example.com".to_owned()]),
-                blocked_domains: BTreeSet::from(["deny.example.com".to_owned()]),
-                timeout_seconds: Some(5),
-                max_bytes: Some(4_096),
-                max_redirects: Some(2),
-            },
-        };
-        let execution = ConstrainedSubagentExecution {
-            mode: crate::conversation::ConstrainedSubagentMode::Async,
-            isolation: crate::conversation::ConstrainedSubagentIsolation::Shared,
-            owner_kind: None,
-            depth: 1,
-            max_depth: 2,
-            active_children: 0,
-            max_active_children: 3,
-            timeout_seconds: 60,
-            allow_shell_in_child: false,
-            child_tool_allowlist: vec!["web.fetch".to_owned()],
-            workspace_root: None,
-            runtime_narrowing: narrowing,
-            kernel_bound: false,
-            identity: Some(ConstrainedSubagentIdentity {
-                nickname: Some("child-research".to_owned()),
-                specialization: Some("researcher".to_owned()),
-            }),
-            profile: Some(ConstrainedSubagentProfile::for_child_depth(1, 2)),
-        };
-
-        let contract = execution.contract_view();
-        let summary = base
-            .delegate_child_prompt_summary(Some(&contract))
-            .expect("delegate child prompt summary");
-
-        assert_eq!(
-            summary,
-            "[delegate_child_runtime_contract]\n\
-Plan within these child-session runtime limits:\n\
-- subagent mode: async\n\
-- subagent nickname: child-research\n\
-- subagent specialization: researcher\n\
-- subagent depth budget: 1/2\n\
-- subagent active-child budget snapshot: 0/3\n\
-- child timeout seconds: 60\n\
-- child bash: denied\n\
-- child tool allowlist: web\n\
-- child runtime binding: direct\n\
-- subagent role: orchestrator\n\
-- subagent control scope: children\n\
-- web network private hosts: denied\n\
-- web network allowed domains: none (effective intersection is empty)\n\
-- web network blocked domains: base-block.example.com, deny.example.com\n\
-- web network timeout seconds: 3\n\
-- web network max bytes: 2048\n\
-- web network max redirects: 1\n\
-- browser max sessions: 1\n\
-- browser max links: 4\n\
-- browser max text chars: 512\n\
-Treat these as enforced limits for this child session."
-        );
-    }
-
-    #[test]
-    fn delegate_child_prompt_summary_omits_disabled_web_fetch() {
-        let base = ToolRuntimeConfig {
-            web_fetch: WebFetchRuntimePolicy {
-                enabled: false,
-                allow_private_hosts: true,
-                enforce_allowed_domains: false,
-                allowed_domains: BTreeSet::new(),
-                blocked_domains: BTreeSet::new(),
-                timeout_seconds: 30,
-                max_bytes: 1_048_576,
-                max_redirects: 5,
-            },
-            browser: BrowserRuntimePolicy {
-                enabled: true,
-                max_sessions: 4,
-                max_links: 16,
-                max_text_chars: 8_192,
-            },
-            ..ToolRuntimeConfig::default()
-        };
-        let narrowing = ToolRuntimeNarrowing {
-            web_fetch: WebFetchRuntimeNarrowing {
-                allow_private_hosts: Some(false),
-                timeout_seconds: Some(5),
-                ..WebFetchRuntimeNarrowing::default()
-            },
-            browser: BrowserRuntimeNarrowing {
-                max_sessions: Some(2),
-                ..BrowserRuntimeNarrowing::default()
-            },
-        };
-
-        let contract = ConstrainedSubagentContractView::from_runtime_narrowing(narrowing);
-        let summary = base
-            .delegate_child_prompt_summary(Some(&contract))
-            .expect("should still render browser section");
-
-        assert!(
-            !summary.contains("web.fetch"),
-            "disabled web_fetch fields should not appear in prompt summary: {summary}"
-        );
-        assert!(
-            summary.contains("- browser max sessions: 2"),
-            "enabled browser fields should still appear: {summary}"
-        );
-    }
-
-    #[test]
-    fn delegate_child_prompt_summary_omits_disabled_browser() {
-        let base = ToolRuntimeConfig {
-            web_fetch: WebFetchRuntimePolicy {
-                enabled: true,
-                allow_private_hosts: true,
-                enforce_allowed_domains: false,
-                allowed_domains: BTreeSet::new(),
-                blocked_domains: BTreeSet::new(),
-                timeout_seconds: 30,
-                max_bytes: 1_048_576,
-                max_redirects: 5,
-            },
-            browser: BrowserRuntimePolicy {
-                enabled: false,
-                max_sessions: 4,
-                max_links: 16,
-                max_text_chars: 8_192,
-            },
-            ..ToolRuntimeConfig::default()
-        };
-        let narrowing = ToolRuntimeNarrowing {
-            web_fetch: WebFetchRuntimeNarrowing {
-                timeout_seconds: Some(5),
-                ..WebFetchRuntimeNarrowing::default()
-            },
-            browser: BrowserRuntimeNarrowing {
-                max_sessions: Some(2),
-                max_links: Some(8),
-                ..BrowserRuntimeNarrowing::default()
-            },
-        };
-
-        let contract = ConstrainedSubagentContractView::from_runtime_narrowing(narrowing);
-        let summary = base
-            .delegate_child_prompt_summary(Some(&contract))
-            .expect("should still render web_fetch section");
-
-        assert!(
-            !summary.contains("browser"),
-            "disabled browser fields should not appear in prompt summary: {summary}"
-        );
-        assert!(
-            summary.contains("- web network timeout seconds: 5"),
-            "enabled web_fetch fields should still appear: {summary}"
-        );
-    }
-
-    #[test]
-    fn delegate_child_prompt_summary_returns_none_when_all_tools_disabled() {
-        let base = ToolRuntimeConfig {
-            web_fetch: WebFetchRuntimePolicy {
-                enabled: false,
-                ..WebFetchRuntimePolicy::default()
-            },
-            browser: BrowserRuntimePolicy {
-                enabled: false,
-                ..BrowserRuntimePolicy::default()
-            },
-            ..ToolRuntimeConfig::default()
-        };
-        let narrowing = ToolRuntimeNarrowing {
-            web_fetch: WebFetchRuntimeNarrowing {
-                timeout_seconds: Some(5),
-                ..WebFetchRuntimeNarrowing::default()
-            },
-            browser: BrowserRuntimeNarrowing {
-                max_sessions: Some(2),
-                ..BrowserRuntimeNarrowing::default()
-            },
-        };
-        let contract = ConstrainedSubagentContractView::from_runtime_narrowing(narrowing);
-
-        assert_eq!(
-            base.delegate_child_prompt_summary(Some(&contract)),
-            None,
-            "should return None when all narrowed tools are disabled"
-        );
-    }
-
-    #[test]
-    fn delegate_child_prompt_summary_renders_profile_even_when_narrowing_is_empty() {
-        let contract = ConstrainedSubagentContractView::from_profile(
-            ConstrainedSubagentProfile::for_child_depth(1, 1),
-        );
-        let summary = ToolRuntimeConfig::default()
-            .delegate_child_prompt_summary(Some(&contract))
-            .expect("profile-only child prompt summary");
-
-        assert_eq!(
-            summary,
-            "[delegate_child_runtime_contract]\n\
-Plan within these child-session runtime limits:\n\
-- subagent role: leaf\n\
-- subagent control scope: none\n\
-Treat these as enforced limits for this child session."
-        );
-    }
-
     #[cfg(feature = "feishu-integration")]
     #[test]
     fn from_env_enables_feishu_runtime_when_credentials_exist() {
@@ -3181,3 +2752,7 @@ Treat these as enforced limits for this child session."
         );
     }
 }
+
+#[cfg(test)]
+#[path = "runtime_config_delegate_prompt_tests.rs"]
+mod delegate_prompt_tests;
